requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22
itsdangerous==2.2.0
google-cloud-documentai==3.11.0
google-cloud-storage==3.10.0
google-cloud-translate==3.24.0
azure-storage-blob==12.28.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.31.0
cloud-sql-python-connector[asyncpg]==1.16.0
